Online.

+++ dingtalk/csharp/core/workbench_1_0/Models/QueryShortcutScopesResponseBody.cs
@@ -18,7 +18,7 @@ namespace AlibabaCloud.SDK.Dingtalkworkbench_1_0.Models
 
         [NameInMap("deptVisibleScopes")]
         [Validation(Required=false)]
-        public List<float?> DeptVisibleScopes { get; set; }
+        public List<long?> DeptVisibleScopes { get; set; }
 
     }
